pyproject.toml
[project]
name = "ps3-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "pymongo>=4.6.0",
    "minio>=7.2.0",
    "pika>=1.3.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.1.1",
]